requirements.txt
django
djangorestframework
psycopg2-binary
uvicorn
redis
celery
drf-yasg